// src/types/contracts/logistic/logistic.ts
export type TMapViewConfigData = {
  deliveryLocationMap: Map<any, any>;
  pickupLocationMap: Map<any, any>;
  qualityTypesMap: Map<any, any>;
  regionMap: Map<any, any>;
  vehicleMap: Map<any, any>;
  goodsMap: Map<any, any>;
  shippingsMap: Map<any, any>;
  purchaseContractsMap: Map<any, any>;
};